Control assembleCoadd warp type with config `warpType`
* For makeCoaddTempExp, use makeDirect/makePsfMatched.
* For assembleCoadd, use warpType.
Deprecated doPsfMatch continues to work for both.

makeDirect/makePsfMatched configs are not appropriate for assembleCoadd
because an instance can only generate a coadd of one warp type.
